requirements.txt
google-generativeai
python-dotenv
pathlib
flask
gunicorn
flask_cors